Add files via upload
The do-file '0_run_dofiles' executes all other do-file codes. To do this, you need to adjust the position of the data in your directory according to the folder parameters set there. 

It is important that you check how the information you download from the DANE repository comes in, because there are usually changes in the type of files, so you need to adjust the '1_create_anual_files' do-file according to the characteristics of the raw data you download from there.
 
The codes shared here are made to replicate from the creation of the database until the final results are obtained.